requirements.txt
django
django-debug-toolbar
django-extensions
django-admin-rangefilter
django-import-export
django-silk
django-storages
django-crispy-forms
django-filter
django-tables2
djangorestframework
djangorestframework-yaml
djangorestframework-xml


flake8
flake8-builtins
flake8-import-order
flake8-print
flake8-breakpoint

boto3
ipython
python-decouple
celery
beautifulsoup4

#pytest
pytest
pytest-django
pytest-cov
pytest-mock